+++ indico/modules/logs/models/entries.py
@@ -47,6 +47,12 @@ class UserLogRealm(RichIntEnum):
     acl = 3
 
 
+class AppLogRealm(RichIntEnum):
+    __titles__ = (None, _('System'), _('Admin'))
+    system = 1
+    admin = 2
+
+
 class LogKind(IndicoIntEnum):
     other = 1
     positive = 2
@@ -259,3 +265,26 @@ class UserLogEntry(LogEntryBase):
         ),
         foreign_keys=[target_user_id]
     )
+
+
+class AppLogEntry(LogEntryBase):
+    """Log entries for the application."""
+
+    __auto_table_args = {'schema': 'indico'}
+    user_backref_name = 'app_log_entries'
+
+    #: The general area of the application the entry comes from
+    realm = db.Column(
+        PyIntEnum(AppLogRealm),
+        nullable=False
+    )
+
+    @staticmethod
+    def log(realm, kind, module, summary, user=None, type_='simple', data=None, meta=None):
+        entry = AppLogEntry(user=user, realm=realm, kind=kind, module=module, type=type_, summary=summary,
+                            data=(data or {}), meta=(meta or {}))
+        db.session.add(entry)
+        return entry
+
+    def __repr__(self):
+        return format_repr(self, 'id', 'logged_dt', 'realm', 'module', _text=self.summary)